+++ TaskGame/Scene/SceneManager.cpp
@@ -33,7 +33,7 @@ void SceneManager::Update(const InputState& input)
 
 void SceneManager::Draw()
 {
-	for (int i = m_scene.size() - 1; i >= 0; --i)
+	for (int i = static_cast<int>(m_scene.size()) - 1; i >= 0; --i)
 	{
 		m_scene[i]->Draw();
 	}
